Consolidate arg length validation
First, we add arg length validation to CreateDocEntryForInstallMethod.

Next, we remove arg length validation from the various
Declare*WithDocumentation functions, as it is superfluous there: If the
arg list is too short, this simply triggers a GAP out of bounds error,
which should be sufficient for debugging. And beyond that, the
CreateDocEntry* functions already perform arg length validation.

This reduces code duplication, with all the usual benefits.