Revise computation of labels
Instead of relying on random numbers, we use a hash of the objects's name,
plus a label counter (which helps to avoid giving things with the same name
the same label).

As hash function we use CRC32, which has the advantage of already being
provided by GAP (and this is what GAPDoc uses, too). Drawback: This is
not a very good hash and collisions are possible. But for our purposes
it should be fine.