doc: major overhaul of the documentation
This removes most of the repetition in the manual, which makes maintenance
easier, and also helps user who now have a much easier time to spot the
differences between the various AutoDoc functions.

The readability of both the documentation source as well as of the
generated manual is also improved.