magic: derive bookname from package_info.PackageDoc.BookName
If this is NOT present, then warn the user about this and suggest
how to correct this.

Note that it is not anymore possible to specify the bookname via a
parameter to AutoDoc(). Indeed, if PackageDoc.BookName is missing,
then GAP will not see the manual anyway, rendering the bookname
irrelevant in this case.