Handle absolute directories for doc_dir in AutoDoc()
Allow AutoDoc() to succeed with a dir entry in the options record which
consists of an IsDirectory() object with an absolute path. There are two steps
to this change: First, some effort must be made to correctly compute
doc_dir_rel under such a circumstance. This commit uses the approach that if
the pkgdir is a prefix of the doc_dir, then doc_dir_rel is the remainder of
the doc_dir when the pkgdir is eliminated.

If doc_dir is absolute and pkgdir is not a prefix thereof, then there is no
sensible value for doc_dir_rel. Therefore, the second step of this change must
be to handle downstream the case that doc_dir_rel is not set. This commit
attempts to do so by presuming in such a case that there is no way to convert
the gapdoc.files to paths relative to the doc_dir, and instead converts them
to absolute paths. It may be the case that this strategy produces non-portable
documentation files, but at least MakeGAPDocDoc succeeds, and it seems to be
the only alternative in such case.

Finally, in my previous experience contributing to GAP there was a strong
preference that all changes include tests for the new behavior. As there was
no test harness for AutoDoc, this commit also adds a boilerplate tst/testall.g
and a single new test, tst/worksheet.tst, which invokes AutoDoc with an
absolutely specified doc_dir.