Make @BeginExampleSession/@EndExampleSession respect plain_text_mode
In order to uniformize the handling of `#!` prefixes in AutoDoc files,
gap/Parser.gi is changed to pass lines between @BeginExampleSession and
@EndExampleSession through unchanged when plain_text_mode is true.

Note this commit also supplies a test file tst/plaintextmode.tst; that file
will not have any effect unless/until a test harness such as the one supplied
with pull request #166 is put in place. But it does at least serve as a
self-documenting illustration of the intended behavior post this change, and
should otherwise be innocuous.