Allow AutoDoc() to take absolute dirs and run from any dir
Allow AutoDoc() to succeed with a dir entry in the options record which
consists of an IsDirectory() object with an absolute path. There are two steps
to this change: First, some effort must be made to correctly compute
doc_dir_rel under such a circumstance. This commit uses the approach that if
the pkgdir is a prefix of the doc_dir, then doc_dir_rel is the remainder of
the doc_dir when the pkgdir is eliminated.

If doc_dir is absolute and pkgdir is not a prefix thereof, then there is no
sensible value for doc_dir_rel. Therefore, the second step of this change must
be to handle downstream the case that doc_dir_rel is not set. This commit
attempts to do so by presuming in such a case that there is no way to convert
the gapdoc.files to paths relative to the doc_dir, and instead converts them
to absolute paths. It may be the case that this strategy produces non-portable
documentation files, but at least MakeGAPDocDoc succeeds, and it seems to be
the only alternative in such case.

Finally, to address the need to test modifications to AutoDoc, this commit adds
a tst/testall.g that compares the results of running AutoDoc on some worksheets
with reference files for those worksheets. It also invokes a boilerplate
TestDirectory() call for maximum flexibility in adding future tests.

Those ingredients created the opportunity to test the generation of the AutoDoc
manual itself, and so this commit adds another new test, tst/dogfood.tst, which
compares the result of AutoDoc on itself with reference files.

To support this self-test, the commit also had to change makedoc.g to allow
the parameter record to be extracted by the test, and it had to allow AutoDoc
itself to be run from an arbitrary directory, not just the package directory of
the package being documented.

A minor point: this commit changes the Print() of the output directory to an
Info() call so that it can be suppressed, to ensure that test results will be
independent of the directory structure on the testing machine.

Resolves: #165.
In addition, the testing portion of this pull request seems to contribute some
progress toward #152 and/or #57.