Add extract_examples option
If this option is passed to AutoDoc(), then all examples in the manual
are extracted into .tst files, which are placed into the tst/ directory
of the package.

This way, all manual tests can be run via the same `TestDirectory` invocation
used for all other tests. And the test can be regenerated from the manual
as part of a package's `makedoc.g` script, which in turn means they are
regenerated when a package release is made using ReleaseTools.