Recognize extract_examples as option to AutoDoc
There are two ways to pass options to the AutoDoc function:

(1) Using the "options stack":
      AutoDoc( package : options )

(2) Using a second argument which is a record:
      AutoDoc( package, options )

All the options that AutoDoc expect can be passed in either of these
two forms, except the option extract_examples, which is only accepted
in form (2).  I changed this so that extract_examples is accepted in
both forms.

(It seems like form (1) is undocumented.  An alternative approach to
fix this inconsistency -- and avoid similar inconsistencies in the
future -- would be to remove form (1) completely.)