Handle documentation of DeclareCategoryCollection declarations.
Note that AutoDoc here duplicates logic in gap/lib/coll.gd to determine the
actual name of the symbol being defined by the call (since that symbol is
computed on the fly by the declaration and does not appear verbatim anywhere
in the call). This is unfortunate, but I could not see any way to call the
relevant code in the standard library, as the name computation of the
collections category is not split out into a separately executable function.

Also, modify worksheet.g to include an example of DeclareCategoryCollection to
make sure the new capacity is tested.