Make @Begin.../@EndExampleSession respect plain_text_mode
In order to uniformize the handling of #! prefixes in AutoDoc files,
gap/Parser.gi is changed to pass lines between @BeginExampleSession and
@EndExampleSession through unchanged when plain_text_mode is true.

Note this commit also supplies two test worksheets. They will not have any
effect unless/until the test harness supplied with pull request #180 is put in
place. But the worksheets do at least serve as illustrations of the intended
behavior post this change, and should otherwise be innocuous.

Co-Authored-By: Max Horn <max@quendi.de>