CreateMainPage: refactor the entity handling
As a side effect, this fixes the &see; entity, which was broken for some time:
the LaTeX `\to` in it was inserted into the XML as a tab followed by `o`.
Luckily I made a mistake while refactoring, which lead to a diff in the test
files, and it immediately became clear that both the new and the old output
were wrong... Now they agree and are both correct. :-)