Remove duplicate entries in autodoc.files
Application: I have multiple .autodoc files which should be detected via
"scan_dirs". However, there is a main .autodoc file which should be
included first as it defines the order of chapters and sections. Thus, I
pass this file to "files" explicitly. This commit makes sure that it is not
included twice if I do that.

Note: Since the order is relevant I have used DuplicateFreeList instead of
Set to keep the order of the first occurrence.