Output entities into _entities.xml
... which is always generated and is included by the main scaffold page. The
main advantage of this approach is that it allows using entities defined
this way even when using a custom main page.

Note that the generated data is not really XML, so the file extension is
misleading. However, the chosen filename has the advantage that it fits
with the names for most other files generated by AutoDoc, and hence it is
already matched by existing .gitignore rules.